% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calKaKs.R
\name{calKaKs}
\alias{calKaKs}
\title{calKaKs}
\usage{
calKaKs(maf, vaf_cutoff = 0.05, class = "SP", parallel = TRUE)
}
\arguments{
\item{maf:}{maf data}

\item{VAF_cutoff:}{Removing mutations with low variant allele frequency (VAF).}

\item{class:}{The class which would be represented.}

\item{parallel:}{Whether calculate the Ka/Ks in parallel? Default: TRUE
"SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS". see MesKit classifyMut.}
}
\description{
calculate the Ka/Ks of each group.
}
