% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRoutines.R
\name{calRoutines}
\alias{calRoutines}
\title{calRoutines}
\usage{
calRoutines(
  maf,
  PrimaryId,
  patient.id = NULL,
  CCF_cutoff = 0.1,
  pairByTumor = TRUE,
  use.tumorSampleLabel = FALSE,
  maf_drivers = NULL,
  subtitle = "both",
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link[MesKit]{readMaf}} function.}

\item{PrimaryId}{Primary tumor IDs to indicate the primary-metastases relationships.}

\item{patient.id}{Select the specific patients. Default NULL, all patients are included.}

\item{CCF_cutoff}{The minimal cutoffs for the present status. The mutations with CCF smaller than CCF_cutoff were consider as absent statuses.}

\item{pairByTumor}{Compare JSI between different tumors. Default TRUE.}

\item{use.tumorSampleLabel}{Logical (Default: FALSE). Rename the 'Tumor_Sample_Barcode' by 'Tumor_Sample_Label'.}

\item{maf_drivers}{Driver information. Two columns are required, including "Mut_ID" and "is.driver". If provided, add the driver mutations in the plot. "Mut_ID" = str_c(Chromosome, Start_Position, Reference_Allele , Tumor_Seq_Allele2, sep = ":")}

\item{subtitle}{the information shows in the subtitle. Options including "JSI", "Hindex", "both" and "none". Default is "both"}

\item{...}{Other options passed to \code{\link[MesKit]{subMaf}}}
}
\description{
We calculate the H index and JSI index according to the pair-wise comparison of CCF.
}
\examples{
cal <- calRoutines(
  maf = maf,
  patient.id = "Met1",
  PrimaryId = "Coad",
  pairByTumor = TRUE,
  use.tumorSampleLabel = TRUE,
  subtitle = "both"
)

plot_grid(plotlist = cal$Met1$plist, nrow = 1)
## add driver information
maf_driver <- data.frame(
  Mut_ID = c("5:112170777:CAGA:-", "1:147092680:-:C"),
  is.driver = c(TRUE, TRUE)
)
cal <- calRoutines(
  maf = maf,
  patient.id = "Met1",
  PrimaryId = "Coad",
  pairByTumor = TRUE,
  use.tumorSampleLabel = TRUE,
  subtitle = "both",
  maf_drivers = maf_driver
)

plot_grid(plotlist = cal$Met1$plist, nrow = 1)
}
\references{
\enumerate{
\item H index. Hu, Z., et al., Quantitative evidence for early metastatic seeding in colorectal cancer. Nat Genet, 2019. 51(7): p. 1113-1122.
\item JSI index. Hu, Z., et al., Multi-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. Nat Genet, 2020. 52(7): p. 701-708.
}
}
