% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinSites.R
\name{getClinSites}
\alias{getClinSites}
\title{getClinSites}
\usage{
getClinSites(maf, Patient_ID = NULL)
}
\arguments{
\item{maf:}{maf data}

\item{Patient_ID:}{Patient_ID. See specific patient.}
}
\description{
We capture the clinical targetable sites of oncokb based on the mutation status.
}
\details{
We match the drivers genes between maf files and clinical sites in oncokb. We only match the gene names, whereas ignoring the cancer types and gene alterations. The main targertable alterations include gene fusions (like BCR-ABL1 fusion), Oncogenic mutations, Exon deletions/insertion, Amplifications, Deletions and Singles-nucleotide mutation (BRAC V600E). Please mannual check the mutation status.

Reference: oncokb.org
}
