% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2variants.R
\name{maf2variants}
\alias{maf2variants}
\title{maf2variants}
\usage{
maf2variants(maf, patient.id = NULL, ccf.cutoff = 0.1, extract.VAF = FALSE)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{readMaf()} function}

\item{patient.id}{Select the specific patients. Default \code{NULL}, all patients are included.}

\item{ccf.cutoff}{Removing low-CCF mutations (default: 0.1).}

\item{extract.VAF}{Whether extract the VAF information. Default \code{FALSE}: extract CCF rather than VAF.}
}
\description{
Change the maf object into variants data frame for the subclonal structures analysis.
}
\details{
This function extracts the \code{Cluster} information from the CCF data. Therefore, the \code{Cluster} column is required in the ccfFile.

For the output \code{variants}, the first five columns are Mutid, Hugo_Symbol, Variant_Classification, Patient_ID, Cluster. The remaining columns indicate variant cellular prevalence for each sample.
}
\examples{
#' data.type <- "split1"

maf1 <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)

ccfs = maf2variants(maf1, patient.id = "Met1")

#extract VAF rather than CCF.
vafs = maf2variants(maf1, patient.id = "Met1", extract.VAF = T)

}
