% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPTevol-package.R
\docType{package}
\name{MPTevol-package}
\alias{MPTevol}
\alias{MPTevol-package}
\title{MPTevol: Clonal Evolutionary History and Metastatic Routines Analysis for Multiple Primary Tumors}
\description{
Provides a practical computation framework for dissecting the evolution of multiple primary tumors (MPT), reducing analysis complexity with modular design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qingjian1991.github.io/MPTevol/}
}

}
\author{
\strong{Maintainer}: Qinjian Chen \email{chenqingjian2010@163.com} (\href{https://orcid.org/xxx}{ORCID})

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})
}

}
\keyword{internal}
