% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNATree.R
\name{plotCNAtree}
\alias{plotCNAtree}
\title{plotCNAtree}
\usage{
plotCNAtree(
  dist,
  bootstrap.rep.num = 500,
  group = NULL,
  group.colors = NULL,
  title = "Cancer"
)
}
\arguments{
\item{dist}{dist files that generated by MEDICC.}

\item{bootstrap.rep.num}{number of bootstrap steps.}

\item{group}{a list that used to indicate the sample groups}

\item{group.colors}{an array indicates the colors of sample groups.}

\item{title}{title of the plot.}
}
\description{
Plot phylogenetic trees of CNAs. The CNAs trees were constructed by MEDICC.
}
\examples{

#read samples distances.
#This dist file is the output of MEDICC
dist = system.file(package="MPTevol", "extdata", "tree_final.dist")

#set group information
group <- list(NORMAL = "NORMAL",
              Breast =  paste0("Breast_", 1:5),
              Coad = paste0("Coad_", 1:5),
              Lung  = paste0("Lung_", 1:5),
              OveryLM = paste0("OveryLM_", 1:5),
              OveryRM = paste0("OveryRM_", 1:6),
              UterusM = paste0("UterusM_", c(1:7))
)

#set group colors
group.colors = setNames( set.colors(n = length(group)), nm = names(group) )

#built trees
tree = plotCNAtree(dist = dist,
                   group = group,
                   group.colors = group.colors
)

tree$plot

}
