% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutTree.R
\name{plotMutTree}
\alias{plotMutTree}
\title{plotMutTree}
\usage{
plotMutTree(
  maf,
  patient.id,
  method = "NJ",
  min.vaf = 0.02,
  bootstrap.rep.num = 500,
  group = NULL,
  group.colors = NULL,
  title = "cancer",
  hexpand_ratio = 0.3,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function}

\item{patient.id}{Select the specific patients.}

\item{method}{Approach to construct phylogenetic trees. Choose one of "NJ"(Neibor-Joining), "MP"(maximum parsimony), "ML"(maximum likelihood), "FASTME.ols" or "FASTME.bal".}

\item{min.vaf}{The minimum value of vaf. Default 0.}

\item{bootstrap.rep.num}{Bootstrap iterations. Default 500.}

\item{group}{a list that used to indicate the sample groups}

\item{group.colors}{an array indicates the colors of sample groups.}

\item{title}{title of the plot.}

\item{...}{parameters in the getPhyloTree function.}

\item{hexpand_ratio:}{hexpand ratio. see \code{\link[ggtree]{hexpand}}}
}
\description{
Plot phylogenetic tree of mutations.
}
\examples{

library(MesKit)

# build a maf data: see MesKit

data.type <- "split1"

maf <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)

# construct a group
group <- list(
  Coad = paste0("Coad_", 1:5),
  OveryLM = paste0("OveryLM_", 1:5),
  OveryRM = paste0("OveryRM_", 1:6),
  UterusM = paste0("UterusM_", c(1, 3))
)

plotMutTree(maf, patient.id = "Met1", group = group, title = "CRC Met")
}
