% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVafCluster.R
\name{plotVafCluster}
\alias{plotVafCluster}
\title{plotVafCluster}
\usage{
plotVafCluster(
  variants,
  cluster.col.name = "cluster",
  vaf.col.names,
  clone.colors = NULL,
  violin = FALSE,
  box = TRUE,
  jitter = TRUE,
  founding.cluster = 1,
  output.file = NULL,
  highlight = NULL,
  highlight.note.col.name = NULL
)
}
\arguments{
\item{variants}{data frame of the variants.
At least cluster column and VAF or CCF columns are required.
Cluster column should contain cluster identities as continuous integer values
starting from 1.}

\item{cluster.col.name}{the column names that containing cluster
information (Default = "cluster").}

\item{vaf.col.names}{the column names of samples containing VAF.}

\item{clone.colors}{setting clone colors.}

\item{violin}{whether plotting violin (Default = FALSE).}

\item{box}{whether plotting box (Default = TRUE).}

\item{jitter}{whether plotting jitter plot (Default = TRUE).}

\item{founding.cluster}{the name of founding clones, one of the most important parameters. For most of circumstances, the founding cluster is the cluster with the highest average CCF cluster.}

\item{output.file}{the output file name (Default = NULL)}

\item{highlight}{column name to indicate whether highlight the sites (TRUE or FALSE).}

\item{highlight.note.col.name}{highlight context.}
}
\value{
a ggplot object
}
\description{
Plot variant clustering in each sample by using combination of box,
violin and jitter plots.
}
