% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCNAProfile.R
\name{readCNAProfile}
\alias{readCNAProfile}
\title{Read CNA Profiles}
\usage{
readCNAProfile(maf, seg, Patient_ID = NULL, purity = 1, ref = "hg19")
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{readMaf()} function}

\item{seg}{seg or seglist.}

\item{Patient_ID}{Patient_ID: select the specific patients.
IF not indicate, the input is Maf and seg, or the input is MafList and segList.}

\item{purity}{purity information for each samples.}

\item{ref}{human reference genome version. Default 'hg19'. Optional: 'hg18' or 'hg38'.}
}
\value{
cnaqc.list for cnaqc initiation.
}
\description{
We used a \code{CNAqc} object, containing a set of mutations, CNA calls and tumor purity values.
The CNAqc was used to deal with the allele-specific CNAs.
}
\details{
This code reads the CNA Profiles for each patient.
The tumor names of  maf and seg are required to match each other.
}
