% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSegment.R
\name{splitSegment}
\alias{splitSegment}
\title{splitSegment}
\usage{
splitSegment(
  segfiles,
  sampleid,
  project.names = "tumor",
  out.dir = "data",
  N.baf = 30,
  cnv_min_length = 100000,
  max_CNt = 15,
  minLength = 100000,
  maxCNV = 4,
  medicc.py = "medicc.py",
  python = "python"
)
}
\arguments{
\item{segfiles}{The allele-specific copy number alterations files generated by \strong{sequenza}}

\item{sampleid}{the corresponding sample ids.}

\item{project.names}{the project names used in the output.}

\item{out.dir}{output dir}

\item{N.baf}{quality control for the sequenza output}

\item{cnv_min_length}{quality control for the sequenza output}

\item{max_CNt}{quality control for the sequenza output}

\item{minLength}{output control: the min length of CNVs to output.}

\item{maxCNV}{output control: the max CNV to output. When the raw CNV greater than maxCNV, then its value was set to maxCNV.}

\item{medicc.py}{the position of meidcc.py}

\item{python}{the position of meidcc.py}
}
\description{
split the segment regions into several parts, according to their shared status. The function first obtains the common shared regions across samples. The corresponding A allele and B allele are output as the format requirements of MEDICC.
}
\details{
This function takes the \strong{sequenza} results as the input and outputs the format requirements of MEDICC.
}
