% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2timescape.R
\name{tree2timescape}
\alias{tree2timescape}
\title{tree2timescape}
\usage{
tree2timescape(results, samples = NULL)
}
\arguments{
\item{results}{the clonal trees that generated by \code{\link[=inferClonalTree]{inferClonalTree()}}.}

\item{samples}{the samples to show in the fisher plot.}
}
\description{
This function generates the input of timescape to visual the fisher plot of
clonal evolution by using the results of \code{\link[=inferClonalTree]{inferClonalTree()}}.
}
