% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewTrees.R
\name{viewTrees}
\alias{viewTrees}
\title{viewTrees}
\usage{
viewTrees(
  phyloTree,
  tree.format = "S4",
  normal.node = "NORMAL",
  group = NULL,
  title = "Cancer",
  showBootstrap = TRUE,
  hexpand_ratio = 0.3
)
}
\arguments{
\item{phyloTree}{phyloTree: The tree is in Parenthetic Format. see \code{\link[read.tree]{subMaf}}.}

\item{tree.format}{the format of tree, S4 or list. Default is S4.}

\item{normal.node}{the sample name of normal sample in the tree.}

\item{group}{a list that used to indicate the sample groups}

\item{showBootstrap}{whether showing the bootstrap values. Default is TRUE.}

\item{title:}{title of the plot.}

\item{hexpand_ratio:}{hexpand ratio. see \code{\link[ggtree]{hexpand}}}
}
\value{
a ggtree object
}
\description{
we used the ggtree to visualize the trees.
}
\examples{

#This dist file is the output of MEDICC
dist = system.file(package="MPTevol", "extdata", "tree_final.dist")

#plot CNA trees without colored samples.
plotCNAtree(dist = dist)

#create a list to indicate the sample groups.
grp <- list(NORMAL = "NORMAL",
    Breast =  paste0("Breast_", 1:5),
    Coad = paste0("Coad_", 1:5),
    Lung  = paste0("Lung_", 1:5),
    OveryLM = paste0("OveryLM_", 1:5),
    OveryRM = paste0("OveryRM_", 1:6),
    UterusM = paste0("UterusM_", c(1:7))
)

plotCNAtree(dist = dist, grp = grp)


}
