% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSegment.R
\name{write.fasta}
\alias{write.fasta}
\title{write.fasta}
\usage{
write.fasta(
  merge_A,
  major = "major",
  out.dir = "data",
  project.names = "tumor"
)
}
\arguments{
\item{project.names}{the project names used in the output.}
}
\description{
Prepare the formate of MEDICC input.
}
