% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calKaKs.R
\name{calKaKs}
\alias{calKaKs}
\title{calKaKs}
\usage{
calKaKs(
  maf,
  patient.id = NULL,
  class = "SP",
  classByTumor = FALSE,
  vaf_cutoff = 0.05,
  parallel = TRUE
)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function}

\item{vaf_cutoff:}{Removing mutations with low variant allele frequency (VAF).}

\item{class:}{The class which would be represented. "SP" (Shared pattern: Public/Shared/Private), other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS". see MesKit \strong{classifyMut}.}

\item{parallel:}{Whether calculate the Ka/Ks in parallel? Default: TRUE}
}
\description{
calculate the Ka/Ks of each group. The mutation are classified as \strong{classifyMut}
}
\examples{

# read data
data.type <- "split1"

maf <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)

# calKaKas

kaks <- calKaKs(maf, patient.id = "Breast", class = "SP", parallel = TRUE, vaf_cutoff = 0.05)

kaks <- calKaKs(maf, patient.id = "Breast", class = "CS", parallel = TRUE, vaf_cutoff = 0.05)

kaks <- calKaKs(maf, class = "SP", parallel = TRUE, vaf_cutoff = 0.05)
}
