% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calKaKs.R
\name{calKaKs}
\alias{calKaKs}
\title{calKaKs calculates the Ka/Ks of each group}
\usage{
calKaKs(
  maf,
  patient.id = NULL,
  class = "SP",
  classByTumor = FALSE,
  vaf.cutoff = 0.05,
  parallel = TRUE
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{readMaf()} function}

\item{patient.id}{Select the specific patients. Default \code{NULL}, all patients are included.}

\item{class}{The class which would be represented.
"SP" (Shared pattern: Public/Shared/Private), other options: "CS" (Clonal status: Clonal/Subclonl)
and "SPCS". see \code{\link[MesKit:classifyMut]{MesKit::classifyMut()}}.}

\item{vaf.cutoff}{Removing mutations with low variant allele frequency (VAF).}

\item{parallel}{If \code{TRUE} (default), run in parallel.}
}
\description{
The mutations are classified by \code{classifyMut()} internally.
}
\examples{
library(MesKit)
data.type <- "split1"

maf <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)

# calKaKas
kaks <- calKaKs(maf, patient.id = "Breast", class = "SP", parallel = TRUE, vaf.cutoff = 0.05)
kaks
kaks <- calKaKs(maf, patient.id = "Breast", class = "CS", parallel = TRUE, vaf.cutoff = 0.05)
kaks
kaks <- calKaKs(maf, class = "SP", parallel = TRUE, vaf.cutoff = 0.05)
kaks
}
