% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinSites.R
\name{getClinSites}
\alias{getClinSites}
\title{getClinSites}
\usage{
getClinSites(maf, Patient_ID = NULL)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function}

\item{Patient_ID:}{Patient_ID. See specific patient.}
}
\description{
We capture the clinical targetable sites of oncokb based on the mutation status.
}
\details{
We match the drivers genes between maf files and clinical sites in oncokb. We only match the gene names, whereas ignoring the cancer types and gene alterations. The main targertable alterations include gene fusions (like BCR-ABL1 fusion), Oncogenic mutations, Exon deletions/insertion, Amplifications, Deletions and Singles-nucleotide mutation (BRAC V600E). Please mannual check the mutation status.

Reference: oncokb.org
}
\examples{
library(MesKit)

# build a maf data: see MesKit

data.type <- "split1"

maf <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.\%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)

# see clinical targetable sites
sites <- getClinSites(maf)

# see one patinets
sites <- getClinSites(maf, Patient_ID = "Breast")

# View data
DT::datatable(sites)
}
