% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPTevol-package.R
\docType{package}
\name{MPTevol-package}
\alias{MPTevol}
\alias{MPTevol-package}
\title{MPTevol: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Qinjian Chen \email{xxx@email.com} (\href{https://orcid.org/xxx}{ORCID})

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})
}

}
\keyword{internal}
