% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNAProfile.R
\name{plotCNAProfile}
\alias{plotCNAProfile}
\title{Visualize CNA profile}
\usage{
plotCNAProfile(cnaqc.list, min_length_show = 1e+05)
}
\arguments{
\item{cnaqc.list}{cnaqc.list}

\item{min_length_show}{the minimal length of CNVs to show.}
}
\description{
This function plots the allele-specific CNAs of multiple-samples.
See \code{\link[=readCNAProfile]{readCNAProfile()}} for examples.
}
