% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classS4.R
\docType{class}
\name{Seg-class}
\alias{Seg-class}
\alias{Segment}
\title{Seg Class}
\description{
Seg Class.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table of Seg File containing CNA information.}

\item{\code{sample.inof}}{data.frame of sample information per patient.}

\item{\code{ref.build}}{human reference genome version. Default 'hg19'. Optional: 'hg18' or 'hg38'.}

\item{\code{allele}}{Indicate whether this is allele-specific CNAs. Default: TRUE.}
}}

