% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSegment.R
\name{splitSegment}
\alias{splitSegment}
\title{splitSegment}
\usage{
splitSegment(
  segfiles,
  sampleid,
  out.dir = "data",
  tumor = "tumor",
  N.baf = 30,
  cnv_min_length = 1e+05,
  max_CNt = 15,
  minLength = 1e+05,
  maxCNV = 4
)
}
\arguments{
\item{segfiles:}{The allele-specific copy number alterations files generated by \strong{sequenza}}

\item{sampleid:}{the corresponding sample ids.}

\item{out.dir:}{output dir}

\item{tumor:}{the output prefix}

\item{N.baf:}{quality control for the sequenza output}

\item{cnv_min_length:}{quality control for the sequenza output}

\item{max_CNt:}{quality control for the sequenza output}

\item{minLength:}{output control: the min length of CNVs to output.}

\item{maxCNV:}{output control: the max CNV to output. When the raw CNV greater than maxCNV, then its value was set to maxCNV.}
}
\description{
split the segment regions into several parts, according to their shared status. The function first obtains the common shared regions across samples. The corresponding A allele and B allele are output as the format requirements of MEDICC.
}
\details{
This function takes the \strong{sequenza} results as the input and outputs the format requirements of MEDICC.
}
